refactor(core): identify the reliability condition by column, not a declared owner
The `owner` tag this branch added never decided anything. `ConditionOwner` had a
single member, and the predicate read

    item.owner === owner || item.owner === undefined

which with one possible owner value can never be false — so it reduced exactly to
"numeric condition on this annotation". The tag could not be tightened later
either: requiring it would hide the hand-built conditions whose invisibility is
the bug (#380) this branch exists to fix.

Its stated justification does not hold. The doc comment cited "a condition
restored from a bundle written before the tag existed", but `filterQuery` is
never serialized — it has no reference outside the control-bar directory, and
bundle settings persist only the `eatOverlayEnabled`/`eatConfidenceThreshold`
scalars. The tag could not round-trip through any saved artifact.

So the concept is the column, and the code now says so: `query-owned.ts` becomes
`query-annotation-conditions.ts`, `findOwnedCondition`/`replaceOwnedConditions`
become `findConditionsForAnnotation`/`replaceConditionsForAnnotation`, and the
`owner` field and `ConditionOwner` type are gone from the shared condition type.
This matches the capability's own requirement title, "The reliability filter is
identified by column, not by condition shape".

Behaviour is unchanged: same 1667 core / 371 utils / 167 app tests pass, with no
test rewritten beyond the rename and the dropped tag assertion.
